router connected

+++ data/migrations/01-make_cars_table.js
@@ -4,7 +4,7 @@ exports.up = function (knex) {
     table.string('vin', 13).notNullable().unique()
     table.string('make', 128).notNullable()
     table.string('model', 128).notNullable()
-    table.numeric('mileage').notNullable()
+    table.integer('mileage').notNullable()
     table.string('title', 128).defaultTo('N/A')
     table.string('transmission', 128).defaultTo('N/A')
   })
